+++ src/gwt/src/org/rstudio/studio/client/common/reditor/EditorLanguage.java
@@ -33,7 +33,7 @@ public class EditorLanguage
    public static final EditorLanguage LANG_R = new EditorLanguage(
          "mode/r", true);
    public static final EditorLanguage LANG_RDOC = new EditorLanguage(
-         "mode/rdoc", false);
+         "mode/rdoc", false, false);
    public static final EditorLanguage LANG_TEX = new EditorLanguage(
          "mode/tex", false);
    public static final EditorLanguage LANG_SWEAVE = new EditorLanguage(
@@ -43,55 +43,54 @@ public class EditorLanguage
    public static final EditorLanguage LANG_RMARKDOWN = new EditorLanguage(
          "mode/rmarkdown", true);
    public static final EditorLanguage LANG_DCF = new EditorLanguage(
-         "mode/dcf", false);
+         "mode/dcf", false, false);
    public static final EditorLanguage LANG_MERMAID = new EditorLanguage(
-         "mode/mermaid", false);
+         "mode/mermaid", false, true);
    public static final EditorLanguage LANG_RHTML = new EditorLanguage(
          "mode/rhtml", true);
    public static final EditorLanguage LANG_CPP = new EditorLanguage(
          "mode/c_cpp", true);
+   public static final EditorLanguage LANG_STAN = new EditorLanguage(
+         "mode/stan", false, true);
    
    // Modes borrowed from Ace
    public static final EditorLanguage LANG_PLAIN = new EditorLanguage(
          "ace/mode/text", false);
    public static final EditorLanguage LANG_HTML = new EditorLanguage(
-         "ace/mode/html", false);
+         "ace/mode/html", false, true);
    public static final EditorLanguage LANG_CSS = new EditorLanguage(
-         "ace/mode/css", false);
+         "ace/mode/css", false, true);
    public static final EditorLanguage LANG_JAVASCRIPT = new EditorLanguage(
-         "ace/mode/javascript", false);
+         "ace/mode/javascript", false, true);
    public static final EditorLanguage LANG_PYTHON = new EditorLanguage(
-         "ace/mode/python", false);
+         "ace/mode/python", false, true);
    public static final EditorLanguage LANG_SQL = new EditorLanguage(
-         "ace/mode/sql", false);
+         "ace/mode/sql", false, true);
    public static final EditorLanguage LANG_SH = new EditorLanguage(
-         "ace/mode/sh", false);
+         "ace/mode/sh", false, true);
    public static final EditorLanguage LANG_YAML = new EditorLanguage(
-         "ace/mode/yaml", false);
+         "ace/mode/yaml", false, true);
    public static final EditorLanguage LANG_XML = new EditorLanguage(
-         "ace/mode/xml", false);
-   public static final EditorLanguage LANG_GRAPHVIZ = new EditorLanguage("ace/mode/dot", false);
+         "ace/mode/xml", false, true);
+   public static final EditorLanguage LANG_GRAPHVIZ = new EditorLanguage("ace/mode/dot", false, true);
    
-   public static final EditorLanguage LANG_CLOJURE = new EditorLanguage("ace/mode/clojure", false);
-   public static final EditorLanguage LANG_COFFEE = new EditorLanguage("ace/mode/coffee", false);
-   public static final EditorLanguage LANG_CSHARP = new EditorLanguage("ace/mode/csharp", false);
-   public static final EditorLanguage LANG_GITIGNORE = new EditorLanguage("ace/mode/gitignore", false);
-   public static final EditorLanguage LANG_GO = new EditorLanguage("ace/mode/golang", false);
-   public static final EditorLanguage LANG_GROOVY = new EditorLanguage("ace/mode/groovy", false);
-   public static final EditorLanguage LANG_HASKELL = new EditorLanguage("ace/mode/haskell", false);
-   public static final EditorLanguage LANG_HAXE = new EditorLanguage("ace/mode/haxe", false);
-   public static final EditorLanguage LANG_JAVA = new EditorLanguage("ace/mode/java", false);
-   public static final EditorLanguage LANG_JULIA = new EditorLanguage("ace/mode/julia", false);
-   public static final EditorLanguage LANG_LISP = new EditorLanguage("ace/mode/lisp", false);
-   public static final EditorLanguage LANG_LUA = new EditorLanguage("ace/mode/lua", false);
-   public static final EditorLanguage LANG_MATLAB = new EditorLanguage("ace/mode/matlab", false);
-   public static final EditorLanguage LANG_PERL = new EditorLanguage("ace/mode/perl", false);
-   public static final EditorLanguage LANG_RUBY = new EditorLanguage("ace/mode/ruby", false);
-   public static final EditorLanguage LANG_RUST = new EditorLanguage("ace/mode/rust", false);
-   public static final EditorLanguage LANG_SCALA = new EditorLanguage("ace/mode/scala", false);
- 
-   
-   public static final EditorLanguage LANG_STAN = new EditorLanguage("mode/stan", false);
+   public static final EditorLanguage LANG_CLOJURE = new EditorLanguage("ace/mode/clojure", false, true);
+   public static final EditorLanguage LANG_COFFEE = new EditorLanguage("ace/mode/coffee", false, true);
+   public static final EditorLanguage LANG_CSHARP = new EditorLanguage("ace/mode/csharp", false, true);
+   public static final EditorLanguage LANG_GITIGNORE = new EditorLanguage("ace/mode/gitignore", false, false);
+   public static final EditorLanguage LANG_GO = new EditorLanguage("ace/mode/golang", false, true);
+   public static final EditorLanguage LANG_GROOVY = new EditorLanguage("ace/mode/groovy", false, true);
+   public static final EditorLanguage LANG_HASKELL = new EditorLanguage("ace/mode/haskell", false, true);
+   public static final EditorLanguage LANG_HAXE = new EditorLanguage("ace/mode/haxe", false, true);
+   public static final EditorLanguage LANG_JAVA = new EditorLanguage("ace/mode/java", false, true);
+   public static final EditorLanguage LANG_JULIA = new EditorLanguage("ace/mode/julia", false, true);
+   public static final EditorLanguage LANG_LISP = new EditorLanguage("ace/mode/lisp", false, true);
+   public static final EditorLanguage LANG_LUA = new EditorLanguage("ace/mode/lua", false, true);
+   public static final EditorLanguage LANG_MATLAB = new EditorLanguage("ace/mode/matlab", false, true);
+   public static final EditorLanguage LANG_PERL = new EditorLanguage("ace/mode/perl", false, true);
+   public static final EditorLanguage LANG_RUBY = new EditorLanguage("ace/mode/ruby", false, true);
+   public static final EditorLanguage LANG_RUST = new EditorLanguage("ace/mode/rust", false, true);
+   public static final EditorLanguage LANG_SCALA = new EditorLanguage("ace/mode/scala", false, true);
    
    /**
     *
@@ -106,8 +105,17 @@ public class EditorLanguage
          String parserName,
          boolean useRCompletion)
    {
+      this(parserName, useRCompletion, false);
+   }
+   
+   public EditorLanguage(
+         String parserName,
+         boolean useRCompletion,
+         boolean useAceLanguageTools)
+   {
       parserName_ = parserName;
       useRCompletion_ = useRCompletion;
+      useAceLanguageTools_ = useAceLanguageTools;
    }
 
    public String getParserName()
@@ -119,7 +127,13 @@ public class EditorLanguage
    {
       return useRCompletion_;
    }
+   
+   public boolean useAceLanguageTools()
+   {
+      return useAceLanguageTools_;
+   }
 
    private final String parserName_;
    private final boolean useRCompletion_;
+   private final boolean useAceLanguageTools_;
 }