+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/AceEditorNative.java
@@ -72,6 +72,20 @@ public class AceEditorNative extends JavaScriptObject {
       this.setReadOnly(readOnly);
    }-*/;
    
+   public native final void setCompletionOptions(boolean enabled,
+                                                 boolean snippets,
+                                                 boolean live,
+                                                 int characterThreshold,
+                                                 int delayMilliseconds) /*-{
+      this.setOptions({
+        enableBasicAutocompletion: enabled,
+        enableSnippets: enabled && snippets,
+        enableLiveAutocompletion: enabled && live,
+        completionCharacterThreshold: characterThreshold,
+        completionDelay: delayMilliseconds
+      });
+   }-*/;
+   
    public native final void toggleCommentLines() /*-{
       this.toggleCommentLines();
    }-*/;